Renamed `NBDialog` component to `GetDataDialog`
